+++ scripts/socket.js
@@ -1,7 +1,7 @@
 import { MODULE_ID } from "./state.js";
 import { handleJoinTable, handleLeaveTable, handleStartRound, handlePlayerAction, handleResetTable, handleMarkLogsAsSeen } from "./tavern-actions.js";
 import { showRollToUser, showPublicRollFromData } from "./dice.js";
-import { showVictoryFanfare, showBustFanfare, showCoinFlip, playBumpEffect, showFloatingText, showSkillCutIn, showSkillResult, showPrivateFeedback, showImpactRing, showFullSetBurst } from "./ui/fx.js";
+import { showVictoryFanfare, showBustFanfare, showCoinFlip, playBumpEffect, showFloatingText, showSkillCutIn, showSkillResult, showPrivateFeedback, showImpactRing, showFullSetBurst, showCheatResult, showSkillBanner } from "./ui/fx.js";
 
 export let tavernSocket;
 
@@ -28,6 +28,8 @@ export function setupSockets() {
   tavernSocket.register("showVictoryFanfare", showVictoryFanfare);
   tavernSocket.register("showBustFanfare", showBustFanfare);
   tavernSocket.register("showCoinFlip", showCoinFlip);
+  tavernSocket.register("showCheatResult", showCheatResult);
+  tavernSocket.register("showSkillBanner", showSkillBanner);
   tavernSocket.register("showImpactRing", showImpactRing);
   tavernSocket.register("playBumpEffect", playBumpEffect);
   tavernSocket.register("showFullSetBurst", showFullSetBurst);